Ml_Rest/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
email-validator==2.1.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
PyJWT==2.8.0
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2